Changed playermovement to be character controller, to be less physics based and closer to the original DOOM. Added a fog controller to match the original doom. Added a scrolling texture script. Added the beginnings of Hangar as a testing zone for both ingame and to learn ProBuilder (incomplete).

+++ Assets/Scripts/PlayerMovementController.cs
@@ -10,7 +10,7 @@ public class PlayerMovementController : MonoBehaviour {
 	public Transform cameraMain;
 
 	[Header("Player Physics Variables")]
-	public int maxSpeed;
+	public float maxSpeed;
 	public float timeToTopSpeed;
 	public float accelerationSpeed;
 	public float groundingDistance;
